Extract ClickPipes runtime and finalize cloud CLI split
